refactor: move data_uncertainties into rb utils
